refactor(jans-cedarling): move build_context to it's own file
